requirements.txt
aiosqlite==0.18.0
anyio==3.6.2
black==23.1.0
click==8.1.3
fastapi==0.92.0
graphql-core==3.2.3
h11==0.14.0
httptools==0.5.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pydantic==1.10.5
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.25.0
strawberry-graphql==0.156.4
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
